fix: refine logic
