+++ src/lib/components/resource/ResourceCard.tsx
@@ -7,31 +7,30 @@ interface ResourceCardProps {
   isSelected?: boolean;
   onClick: () => void;
 }
+
 export const ResourceCard = ({
   name,
   amount,
   hasBorder = false,
   isSelected = false,
   onClick,
-}: ResourceCardProps) => {
-  return (
-    <Flex
-      w="full"
-      border={hasBorder ? "1px solid" : "none"}
-      _hover={{ background: "gray.800" }}
-      transition="all .25s ease-in-out"
-      cursor="pointer"
-      borderColor="gray.700"
-      bgColor={isSelected ? "gray.800" : "gray.900"}
-      borderRadius={8}
-      p={3}
-      alignItems="center"
-      onClick={onClick}
-    >
-      <Text className="ellipsis">{name}</Text>
-      <Badge variant={isSelected ? "primary" : "gray"} ml={2}>
-        {amount}
-      </Badge>
-    </Flex>
-  );
-};
+}: ResourceCardProps) => (
+  <Flex
+    w="full"
+    border={hasBorder ? "1px solid" : "none"}
+    _hover={{ background: "gray.800" }}
+    transition="all .25s ease-in-out"
+    cursor="pointer"
+    borderColor="gray.700"
+    bgColor={isSelected ? "gray.800" : "gray.900"}
+    borderRadius={8}
+    p={3}
+    alignItems="center"
+    onClick={onClick}
+  >
+    <Text className="ellipsis">{name}</Text>
+    <Badge variant={isSelected ? "primary" : "gray"} ml={2}>
+      {amount}
+    </Badge>
+  </Flex>
+);
