+++ src/lib/components/resource/index.ts
@@ -1,0 +1,2 @@
+export * from "./ResourceCard";
+export * from "./ResourceDetailCard";
