+++ src/lib/pages/account-details/components/resources/types.ts
@@ -1,0 +1,13 @@
+import type { InternalResource } from "lib/types";
+
+export interface ResourceGroup {
+  group: string;
+  account: string;
+  displayName: string;
+  items: InternalResource[];
+}
+
+export interface ResourceGroupByAccount {
+  owner: string;
+  resources: Record<string, ResourceGroup>;
+}
